Put project MenuSystem by Joel Murielle alias josmiley under git version control.
I did this because:
On project webpage http://www.pygame.org/project/2031
link to website where source code of the project should be,
http://joel-murielle.perso.sfr.fr/
does not work currently.
So I wanted to put the source code of the project to github.com
to be available to all again.
And also I want to fork the project
and modify the project a bit
as I consider some attributes of the project to be bugs.